requirements.txt
streamlit
pynacl
qrcode
reportlab